修改方法: checkitem--->新增自定义异常 异常信息捕获通过log日志

+++ health_service/src/main/java/com/itheima/health/service/Impl/CheckItemsServiceImpl.java
@@ -5,16 +5,19 @@ import com.alibaba.dubbo.config.annotation.Service;
 import com.github.pagehelper.Page;
 import com.github.pagehelper.PageHelper;
 import com.github.pagehelper.StringUtil;
+import com.itheima.health.constant.MessageConstant;
 import com.itheima.health.dao.CheckItemsDao;
 import com.itheima.health.entity.PageResult;
 import com.itheima.health.entity.QueryPageBean;
+import com.itheima.health.entity.Result;
+import com.itheima.health.exception.HealthException;
 import com.itheima.health.pojo.CheckItem;
 import com.itheima.health.service.CheckItemsService;
 import org.springframework.beans.factory.annotation.Autowired;
 
 import java.util.List;
 
-@Service
+@Service(interfaceClass = CheckItemsService.class)
 public class CheckItemsServiceImpl implements CheckItemsService {
 
     @Autowired
@@ -61,6 +64,15 @@ public class CheckItemsServiceImpl implements CheckItemsService {
     @Override
     public Boolean deleteById(Integer id) {
 
+        //增加异常管控
+
+        //1.根据删除的id查看有没有和checkgroup进行关联
+        Integer counts = checkItemsDao.count(id);
+        if (counts > 0){
+            //2.如果有则定义一个异常类 抛出自定义异常
+            throw new HealthException("想要删除的检查项已被检查组使用,请先删除检查组!");
+        }
+
         Integer rows = checkItemsDao.deleteById(id);
         return rows>0;
     }
@@ -72,4 +84,6 @@ public class CheckItemsServiceImpl implements CheckItemsService {
 
         return rows>0;
     }
+
+
 }
